--- internal/tlsstate/doc.go
@@ -1,3 +1,0 @@
-// Package tlsstate performs a TLS handshake over the control channel, and they
-// exchanges keys with the server over this secure channel.
-package tlsstate
